src/requirements.example.txt
cryptography==3.4.7
google-api-python-client==2.5.0
google-auth-httplib2==0.1.0
google-auth==1.30.0
google-cloud-pubsub==2.5.0
google-cloud-iot==2.0.2
grpc-google-iam-v1==0.12.3
pyjwt==2.1.0
paho-mqtt==1.5.1